backend/requirements.txt
asgiref==3.9.1
babel==2.17.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cssselect2==0.8.0
diff-match-patch==20241021
Django==5.2.4
django-constance==4.3.2
django-cors-headers==4.7.0
django-filer==3.3.1
django-import-export==4.3.9
django-js-asset==3.1.2
django-money==3.5.4
django-mptt==0.17.0
django-phonenumber-field==8.1.0
django-polymorphic==4.1.0
django-simple-history==3.10.1
django-unfold==0.63.0
easy-thumbnails==2.10
h11==0.16.0
lxml==6.0.0
phonenumbers==9.0.10
pillow==11.3.0
psycopg2-binary==2.9.10
py-moneyed==3.0
python-dotenv==1.1.1
reportlab==4.4.3
setuptools-git==1.2
sqlparse==0.5.3
svglib==1.5.1
tablib==3.8.0
tinycss2==1.4.0
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
webencodings==0.5.1
